imageCaption/requirements.txt
torch>=2.2.1
transformers>=4.38.2
Pillow>=9.4.0
requests>=2.31.0
urllib3==1.26.6
boto3
elasticsearch